language=requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pymongo>=4.3.3
neo4j>=5.8.0
pydantic>=1.10.7
python-dotenv>=1.0.0
streamlit>=1.22.0
requests>=2.28.2
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0 